@@ -1,12 +1,21 @@
 module.exports = {
-  extends: ['airbnb-typescript', 'prettier'],
-  parserOptions: {
-    project: './tsconfig.json',
-  },
-  env: {
-    browser: true,
-  },
+  root: true,
+  env: { browser: true, es2020: true },
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:react-hooks/recommended',
+    'airbnb-typescript',
+    'prettier',
+  ],
+  ignorePatterns: ['dist', '.eslintrc.cjs'],
+  parser: '@typescript-eslint/parser',
+  plugins: ['react-refresh'],
   rules: {
+    'react-refresh/only-export-components': [
+      'warn',
+      { allowConstantExport: true },
+    ],
     'linebreak-style': 'off',
     '@typescript-eslint/quotes': ['error', 'single'],
     'react/jsx-filename-extension': [1, { extensions: ['.tsx'] }],
