@@ -264,8 +264,9 @@ function Nav() {
                     </a>
                   ) : (
                     <NavLink
-                      activeClassName='navLinkActive'
-                      className='navLink'
+                      className={({ isActive }) =>
+                        isActive ? 'navLink navLinkActive' : 'navLink'
+                      }
                       to={link.path}
                       id={link.id}
                     >
@@ -274,8 +275,9 @@ function Nav() {
                   )
                 ) : (
                   <NavLink
-                    activeClassName='navLinkActive'
-                    className='navLink'
+                    className={({ isActive }) =>
+                      isActive ? 'navLink navLinkActive' : 'navLink'
+                    }
                     to={link.path}
                     id={link.id}
                   >
@@ -297,8 +299,9 @@ function Nav() {
                           }
                         >
                           <NavLink
-                            className='navLink'
-                            activeClassName='navLinkActive'
+                            className={({ isActive }) =>
+                              isActive ? 'navLink navLinkActive' : 'navLink'
+                            }
                             to={{ pathname: subLink.path }}
                           >
                             <p>{subLink.name}</p>
@@ -317,8 +320,9 @@ function Nav() {
                         }
                       >
                         <NavLink
-                          className='navLink'
-                          activeClassName='navLinkActive'
+                          className={({ isActive }) =>
+                            isActive ? 'navLink navLinkActive' : 'navLink'
+                          }
                           to={{ pathname: subLink.path }}
                         >
                           <p>{subLink.name}</p>
